+++ src/plone/app/multilingualindexes/testing.py
@@ -1,5 +1,4 @@
-from plone.app.multilingual.testing import PLONE_APP_MULTILINGUAL_PRESET_FIXTURE  # noqa
-from plone.app.robotframework.testing import REMOTE_LIBRARY_BUNDLE_FIXTURE
+from plone.app.multilingual.testing import PLONE_APP_MULTILINGUAL_PRESET_FIXTURE
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
 from plone.app.testing import PloneWithPackageLayer
@@ -25,10 +24,4 @@ PAMI_INTEGRATION_TESTING = IntegrationTesting(
 
 PAMI_FUNCTIONAL_TESTING = FunctionalTesting(
     bases=(PAMI_FIXTURE,), name="PAMILayer:FunctionalTesting"
-)
-
-
-PAMI_ACCEPTANCE_TESTING = FunctionalTesting(
-    bases=(PAMI_FIXTURE, REMOTE_LIBRARY_BUNDLE_FIXTURE, z2.ZSERVER_FIXTURE),
-    name="PAMILayer:AcceptanceTesting",
 )